backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0  # PostgreSQL async driver
# psycopg2-binary - не нужен для async, используем asyncpg для всего
# psycopg2-binary==2.9.9  # PostgreSQL sync driver (для миграций) - опционально
aiosqlite==0.19.0  # SQLite async driver для разработки

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-telegram-bot==20.7

# Telegram Bot (Aiogram)
aiogram==3.2.0
aiohttp==3.9.1

# Google APIs
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
gspread==5.12.4

# Validation
pydantic==2.5.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1

# Date/Time
python-dateutil==2.8.2
pytz==2023.3

# Logging
loguru==0.7.2

# Testing (опционально)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
